feat(objective): improves edit mode structure
