+++ src/components/Objective/AccordionItem/Item/Button/edit-mode.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const EditMode = () => <p>Edit mode</p>
